Moved the code to wait for the volume creation into the Reddwarf Compute manager. Moved most of the new code into a helper class which handles the instance_id, context, and volume_id variables and keeps too much stuff from being passed around everywhere. It should be easier to unit test too.
Added feature to nova.utils.poll_until so that no condition needs to be specified if the retriever returns a boolean result.

Added code to restart the Compute service with altered flag values.

Altered the guest_initialize_failure test to use Proboscis v1.2 in order to avoid duplicating code or using global variables and got the code down to one class. Added it to the group of tests which run as default.