Fixed the DNS stuff to work again.
Updated the DNS daemon to not stop right after being started. Looks like this didn't get updated when the Nova code for the daemon's changed.

The DNS api has become a bit slower which causes some problems (its not really that slow, but is compared to when I started it and the response was instaneous). Additionally the API was changed so that the "future" object that is called to get the status of a created resource now provides a really nice error message. I updated the code to raise a special exception type with the contents of this error message if it is encountered when polling the created resource.

I changed the RsDnsDriver to poll on DNS entry creation for awhile and log the error if one is found, which should make it easier to see what went wrong in the logs (before figuring this out would require listing jobs in the RS DNS API). I also updated several tests to use poll until and give resources a chance to provision instead of assuming the DNS entry would have been created instantly and failing the test.

I also changed the RsDns client and driver functional tests. They'd just been organized incredibly poorly before and I should have changed them earlier. Hopefully now it'll be easier to track down problems.

I added a Python script to show all the DNS entries available to the driver using the information from the flag file. Before this was done using a test, but that was becoming a bit too confusing.

Finally on a bit of a tangent I added some code to the instance tests to allow them to run without first creating an instance. This doesn't work 100% yet (as you can imagine its hard to test) but I'm sure if we all work together we can polish it up.  Its just there to help the devs so it doesn't need to work 100%.