The VerifyManagerAbortsInstanceWhenGuestInstallFails test was failing when run with the other test because the one right before it had shut off the Compute daemon. Now we had code to wait for the daemon to restart which was checking the logs, but this didn't check for the heartbeat the daemon sends periodically (and apparently not right after it starts up). I added code to wait until the service is up and now the scheduler goes ahead and builds the instance. I guess this hasn't broken elsewhere because my machine is currently running slow due to running two VMs, or maybe its because the DNS stuff is enabled.
The code in the scheduler was broken such that if more than one host got returned it would never be checked- instead, the first host (in a list) got checked and if it didn't meet the criteria the other elements were ignored. Didn't test this but it doesn't break the integration tests and the old logic had to be wrong.